refactor: move config files to src/config directory
