+++ src/pages/posts/[id].tsx
@@ -1,4 +1,4 @@
-import { ACTIVE_THEME } from '../../lib/themeSelector';
+import { ACTIVE_THEME } from '../../config/themeSelector';
 import { ArticleServiceFactory } from '../../core/factories/ArticleServiceFactory';
 import { ArticlePageUsecase } from 'application/usecases/ArticlePageUsecase';
 import dynamic from 'next/dynamic';
